+++ plugins/ai.wingbot.jumpBack/plugin.js
@@ -1,0 +1,21 @@
+/**
+ * @param {import('../../src/Request')} req
+ * @param {import('../../src/Responder')} res
+ * @param {Function} postback
+ */
+module.exports = (req, res, postback) => {
+    let { jumpPluginStack = [] } = {
+        ...req.state,
+        ...res.newState
+    };
+
+    const [current, ...rest] = jumpPluginStack;
+    jumpPluginStack = rest;
+    res.setState({ jumpPluginStack });
+
+    if (current) {
+        postback(current.b);
+        return null;
+    }
+    return undefined;
+};